+++ lib/lib.test.ts
@@ -1,29 +1,30 @@
 import { createPackage } from "./createPackage.js";
+import { lib } from "./lib.js";
 
 const mockExitWithSuccess = vi.fn();
 const mockExitWithError = vi.fn();
-vi.mock("./utils.ts", () => ({
+vi.mock("./utils.js", () => ({
 	exitWithSuccess: () => mockExitWithSuccess(),
 	exitWithError: () => mockExitWithError(),
 }));
 
-const mockGetPackagesFromDirectory = vi.fn();
-vi.mock("./getPackagesFromDirectory", () => ({
-	getPackagesFromDirectory: mockGetPackagesFromDirectory,
+const mockGetPackagesFromDirectory = vi.fn().mockResolvedValue([]);
+vi.mock("./getPackagesFromDirectory.js", () => ({
+	getPackagesFromDirectory: () => mockGetPackagesFromDirectory(),
 }));
 
 describe("lib", () => {
 	describe("no pinned lib", () => {
 		it("exits process with 0", async () => {
-			mockGetPackagesFromDirectory.mockResolvedValueOnce([]);
-			await import("./index");
+			await lib();
 			expect(mockExitWithSuccess).toHaveBeenCalled();
 		});
 
 		it.skip("prints a nice message", async () => {
 			const consoleLogSpy = vi.spyOn(console, "log");
 
-			await import("./index");
+			await lib();
+
 			expect(consoleLogSpy).toHaveBeenCalledWith(
 				"It seems all your dependencies are pinned :)",
 			);
@@ -31,7 +32,7 @@ describe("lib", () => {
 	});
 
 	describe("found pinned lib", () => {
-		it.only("exits process with 1", async () => {
+		it("exits process with 1", async () => {
 			mockGetPackagesFromDirectory.mockReturnValue([
 				createPackage({
 					dir: "/path/to/dir",
